requirements.txt
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
pillow>=10.0.0
python-multipart>=0.0.6
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
sqlite3
